last version minitalk

+++ bonus/minitalk_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minitalk_bonus.h                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: heljary <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: heljary <heljary@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 19:38:26 by heljary           #+#    #+#             */
-/*   Updated: 2025/01/28 19:38:41 by heljary          ###   ########.fr       */
+/*   Updated: 2025/03/21 03:49:14 by heljary          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 # define MINITALK_BONUS_H
 # define _GNU_SOURCE
 # include <signal.h>
-# include <unistd.h>
-# include <stdlib.h>
 # include <stdio.h>
+# include <stdlib.h>
+# include <unistd.h>
 
 void	ft_putnbr(int n);
-void    ft_putstr(char *str);
+void	ft_putstr(char *str);
 void	ft_putchar(char c);
-int     ft_atoi(const char *str);
-void    error_exit(char *msg);
+long	ft_atoi(const char *str);
+void	error_exit(char *msg);
 
 #endif
